myBoard/requirements.txt
# myBoard/requirements.txt

django==4.2.7
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-genai
